Change method of calculation of Most Common Day
Change from a for loop to dataframe groupby with the calendar.day_name
function.  No errors/issues with the origninal code - this was just a
slicker way of determining (and probably faster overall) way of determining
the most common day of week.

Credit to Phil Ribbens Udacity bikeshare project for this approach
https://github.com/philribbens/udacity-bikeshare-project/blob/master/bikeshare.py